userExtractor and test

+++ controllers/blog.js
@@ -7,8 +7,24 @@ blogsRouter.get('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-    await Blog.findByIdAndDelete(request.params.id)
-    response.status(204).end()
+
+    const blog = await Blog.findById(request.params.id)
+
+    console.log('blog', blog)
+
+    console.log('request.user._id.toString()', request.user._id.toString())
+    console.log('blog.user.toString()', blog.user.id.toString())
+
+    if(request.user._id.toString() === blog.user.id.toString()){
+        await Blog.findByIdAndDelete(request.params.id)
+        response.status(204).end()
+    }
+    else{
+        return response.status(401).json({ error: 'Blogs can only be deleted by author' })
+    }
+   
+    
+    
 })
 
 blogsRouter.patch('/:id', async (request, response) => {
@@ -43,7 +59,7 @@ blogsRouter.post('/', (request, response, next) => {
             name: request.user.name,
             id: request.user._id
         }
-      
+
     })
 
     blog.likes = blog.likes === undefined ? 0 : blog.likes
